requirements.txt
pandas==3.0.1
sqlalchemy
psycopg2-binary==2.9.11
python-dotenv
